fix deletion bug

+++ Backend/Controllers/CustomerController.js
@@ -832,16 +832,18 @@ const deleteCustomer = async (req, res) => {
   try {
     const { id } = req.body;
     // Check if there are any orders linked to this customer
-   
-    const deletedCustomer = await Customer.findByIdAndDelete(id);
-    if (!deletedCustomer) {
-      return res.status(404).json({ message: "Customer not found" });
-    }
+ if(!id){
+  return res.status(404).json({ message: "Customer not found" });
+ }
     const linkedOrders = await Order.find({ cusid: id });
     if (linkedOrders.length > 0) {
       return res.status(400).json({ message: "Cannot delete customer: Please delete all orders attached to this customer first." });
     }
-    
+       
+    const deletedCustomer = await Customer.findByIdAndDelete(id);
+    if (!deletedCustomer) {
+      return res.status(404).json({ message: "Customer not found" });
+    }
     res.status(200).json({ message: "Customer deleted successfully" });
   } catch (error) {
     console.log(error)
